build script: Simplify workdir path handling: use home directory directly
Previously, we used `~/android-kernel` and set the workdir variable to that directory.
Now, we simplify by using the home directory directly as "workdir", reducing redundancy.

This change also makes it easier if we want to split the build process into multiple jobs in
the GitHub Actions workflow in the future.